requirements.txt
aiobotocore==0.9.4
aiohttp==3.4.4
aiopg==0.15.0
async-timeout==3.0.0
attrs==18.2.0
botocore==1.10.58
chardet==3.0.4
docutils==0.14
idna==2.7
jmespath==0.9.3
jsonschema==2.6.0
multidict==4.4.0
Pillow==5.2.0
psycopg2==2.7.5
python-dateutil==2.7.3
PyYAML==3.13
six==1.11.0
SQLAlchemy==1.2.11
wrapt==1.10.11
yarl==1.2.6
jsonify==0.5